Moved VL53 i2cbus stuff back to module
